Offload established flows from the Calico forward chain
The rule sits at the top of the Calico forward chain, before the per-workload dispatch chains accept established traffic and out of the shared chains that the kernel rejects for flow offload (only the forward hook is allowed).
